+++ cvat/apps/engine/frame_provider.py
@@ -1,2 +1,2 @@
 # Backward compatibility
-from cvat.apps.engine.media_providers.frame_provider import *  # pylint: disable=unused-import, disable=wildcard-import
+from cvat.apps.engine.media_io.frame_provider import *  # pylint: disable=unused-import, disable=wildcard-import
