+++ cvat/apps/engine/media_io/audio_provider.py
@@ -31,13 +31,13 @@ from cvat.apps.engine.media_extractors import (
     IMediaReader,
     Mp3ChunkWriter,
 )
-from cvat.apps.engine.media_providers.media_chunks import (
+from cvat.apps.engine.media_io.media_chunks import (
     BufferChunkLoader,
     ChunkLoader,
     FileChunkLoader,
     ReaderFactory,
 )
-from cvat.apps.engine.media_providers.media_provider import DataWithMeta, IMediaProvider
+from cvat.apps.engine.media_io.media_provider import DataWithMeta, IMediaProvider
 from cvat.apps.engine.utils import take_by
 from utils.dataset_manifest.utils import MemOpenable
